requirements.txt
bcrypt
ctktooltip
customtkinter
openpyxl
packaging
pandas
pillow
pip-chill
